# mlops_backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLOps Backend API",
    "version": "1.0.0",
    "description": "REST API for the MLOps platform. Manages batch inference jobs, registered ML models, datasets, online serving endpoints, endpoint metrics, and alerting. Follows DRF-style conventions: list endpoints return paginated envelopes, write endpoints accept JSON bodies, all identifiers are string slugs."
  },
  "servers": [
    { "url": "http://127.0.0.1:7319", "description": "Local development server (overridden at runtime via MLOPS_API_URL)" }
  ],
  "paths": {
    "/batch_jobs/": {
      "get": {
        "operationId": "mlops_batch_jobs_list",
        "summary": "List batch inference jobs",
        "description": "Returns a paginated list of BatchJob objects, optionally filtered by status, model, or submission time. Pagination uses page/page_size query parameters. The response envelope includes count (total matching items), next (URL of next page or null), previous (URL of previous page or null), and results (array of BatchJob objects).",
        "tags": ["batch_jobs"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by job status. Enum: PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED.",
            "schema": { "type": "string", "enum": ["PENDING", "RUNNING", "SUCCEEDED", "FAILED", "CANCELLED"] }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "description": "Filter by model identifier. Example: \"m-llama3-8b-ft\".",
            "schema": { "type": "string" }
          },
          {
            "name": "submitted_after",
            "in": "query",
            "required": false,
            "description": "Return only jobs submitted after this ISO 8601 datetime. Example: \"2026-04-20T00:00:00Z\".",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number within the paginated result set. Default: 1.",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Number of results per page. Default: 20. Maximum: 100.",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of batch jobs.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedBatchJobList" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "mlops_batch_jobs_create",
        "summary": "Submit a new batch inference job",
        "description": "Creates and queues a new batch inference job. The job starts in PENDING status. model_id and dataset_id are required. shards_total controls how many parallel shards the job is split into (default 16).",
        "tags": ["batch_jobs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchJobCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Batch job created with status PENDING.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchJob" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/batch_jobs/{job_id}/": {
      "get": {
        "operationId": "mlops_batch_jobs_retrieve",
        "summary": "Retrieve a single batch inference job",
        "description": "Returns the BatchJob object for the given job_id. Returns 404 if the job does not exist.",
        "tags": ["batch_jobs"],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "description": "Unique identifier for the BatchJob. Example: \"bj-0001\".",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "BatchJob object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchJob" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "mlops_batch_jobs_destroy",
        "summary": "Cancel and delete a batch job",
        "description": "Cancels the running job and marks it CANCELLED. Returns the job_id and final status. Returns 404 if the job does not exist.",
        "tags": ["batch_jobs"],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "description": "Unique identifier for the BatchJob.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchJobCancelled" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/batch_jobs/{job_id}/logs/": {
      "get": {
        "operationId": "mlops_batch_jobs_logs_retrieve",
        "summary": "Retrieve recent log lines for a batch job",
        "description": "Tails the most recent log lines for the given batch job. The lines parameter controls how many log lines are returned (default 50, maximum 500). Returns the log content as a newline-separated string in the content field.",
        "tags": ["batch_jobs"],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "description": "Unique identifier for the BatchJob.",
            "schema": { "type": "string" }
          },
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "description": "Number of log tail lines to return. Default: 50. Maximum: 500.",
            "schema": { "type": "integer", "default": 50, "maximum": 500 }
          }
        ],
        "responses": {
          "200": {
            "description": "Log content for the batch job.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobLogs" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/batch_jobs/{job_id}/retry_shards/": {
      "post": {
        "operationId": "mlops_batch_jobs_retry_shards_create",
        "summary": "Retry only the failed shards of a batch job",
        "description": "Creates a new child job containing only the shards that failed in the original job. Returns 400 if the job has no failed shards. Returns 404 if the job does not exist. The new_job_id field in the response is the identifier of the newly created retry job.",
        "tags": ["batch_jobs"],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the parent BatchJob.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": {
            "description": "Retry job created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RetryShards" }
              }
            }
          },
          "400": {
            "description": "Job has no failed shards.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/models/": {
      "get": {
        "operationId": "mlops_models_list",
        "summary": "List all registered ML models",
        "description": "Returns a paginated list of all registered Model objects. Use page and page_size to navigate results. Each Model record includes model_id, name, version, framework, status, and created_at timestamp.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number. Default: 1.",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Results per page. Default: 20.",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of models.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedModelList" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "mlops_models_create",
        "summary": "Register a new ML model",
        "description": "Registers a new Model in the registry. name and version are required. The returned object has status \"registered\". Returns 400 if required fields are missing.",
        "tags": ["models"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModelCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Model registered.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Model" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/": {
      "get": {
        "operationId": "mlops_models_retrieve",
        "summary": "Retrieve a single ML model",
        "description": "Returns the Model object for the given model_id. Returns 404 if the model does not exist.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "description": "Unique model identifier. Example: \"m-llama3-8b-ft\".",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Model object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Model" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "mlops_models_partial_update",
        "summary": "Partially update a model record",
        "description": "Applies a partial update (PATCH) to a model. Only the fields provided in the request body are updated. Updatable fields: status (enum: active, deprecated, archived) and name. Returns the list of updated fields. Returns 404 if the model does not exist.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "description": "Unique model identifier.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModelPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelUpdateResult" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/datasets/": {
      "get": {
        "operationId": "mlops_datasets_list",
        "summary": "List all registered datasets",
        "description": "Returns a paginated list of all registered Dataset objects. Each Dataset record includes dataset_id, name, row_count, format, and created_at timestamp. Use page and page_size to navigate results.",
        "tags": ["datasets"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number. Default: 1.",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Results per page. Default: 20.",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of datasets.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedDatasetList" }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_id}/": {
      "get": {
        "operationId": "mlops_datasets_retrieve",
        "summary": "Retrieve a single dataset",
        "description": "Returns the Dataset object for the given dataset_id. Returns 404 if the dataset does not exist.",
        "tags": ["datasets"],
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "description": "Unique dataset identifier. Example: \"ds-support-tickets\".",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dataset object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dataset" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
    "/endpoints/": {
      "get": {
        "operationId": "mlops_endpoints_list",
        "summary": "List online serving endpoints",
        "description": "Returns a paginated list of Endpoint objects, optionally filtered by status. Each Endpoint record describes an online model serving endpoint with its model_id, environment, traffic_splits, and current status. Status enum: ACTIVE, DEGRADED, INACTIVE, PENDING.",
        "tags": ["endpoints"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by endpoint status. Enum: ACTIVE, DEGRADED, INACTIVE, PENDING.",
            "schema": { "type": "string", "enum": ["ACTIVE", "DEGRADED", "INACTIVE", "PENDING"] }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number. Default: 1.",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Results per page. Default: 20.",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of endpoints.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedEndpointList" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "mlops_endpoints_create",
        "summary": "Deploy a new serving endpoint",
        "description": "Deploys a new online serving endpoint for the given model. name and model_id are required. The returned endpoint has status PENDING while it initialises. Returns 400 if required fields are missing.",
        "tags": ["endpoints"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EndpointCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Endpoint created with status PENDING.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Endpoint" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
...
    "/endpoints/{endpoint_id}/logs/": {
      "get": {
        "operationId": "mlops_endpoints_logs_retrieve",
        "summary": "Tail recent request logs for an endpoint",
        "description": "Returns the most recent request log lines for the given endpoint. The lines parameter controls how many lines are returned (default 50). The content field is a newline-separated string of log entries.",
        "tags": ["endpoints"],
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "description": "Unique endpoint identifier.",
            "schema": { "type": "string" }
          },
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "description": "Number of log tail lines. Default: 50.",
            "schema": { "type": "integer", "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Log content for the endpoint.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EndpointLogs" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
...
    "/deploys/{deploy_id}/": {
      "get": {
        "operationId": "mlops_deploys_retrieve",
        "summary": "Retrieve a single deploy",
        "description": "Returns the Deploy object for the given deploy_id. Returns 404 if the deploy does not exist.",
        "tags": ["deploys"],
        "parameters": [
          {
            "name": "deploy_id",
            "in": "path",
            "required": true,
            "description": "Unique deploy identifier. Example: \"dep-001\".",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deploy object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Deploy" }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchJobCreate": {
        "type": "object",
        "description": "Request body for creating a new batch inference job.",
        "properties": {
          "model_id": { "type": "string", "description": "Model to run inference with." },
          "dataset_id": { "type": "string", "description": "Dataset to run inference over." },
          "shards_total": { "type": "integer", "default": 16, "description": "Number of shards to split the job into. Default: 16." }
        },
        "required": ["model_id", "dataset_id"]
      },
      "BatchJobCancelled": {
        "type": "object",
        "description": "Response after cancelling a batch job.",
        "properties": {
          "job_id": { "type": "string" },
          "status": { "type": "string", "example": "CANCELLED" }
        },
        "required": ["job_id", "status"]
      },
...
      "ModelPatch": {
        "type": "object",
        "description": "Request body for partially updating a model record.",
        "properties": {
          "status": { "type": "string", "enum": ["active", "deprecated", "archived"], "description": "New model status." },
          "name": { "type": "string", "description": "New human-readable name." }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "description": "Standard error response.",
        "properties": {
          "detail": { "type": "string", "description": "Human-readable error message." }
        },
        "required": ["detail"]
      },
      "PaginatedBatchJobList": {
        "type": "object",
        "description": "Paginated list of BatchJob objects.",
        "properties": {
          "count": { "type": "integer", "description": "Total number of matching items." },
          "next": { "type": "string", "nullable": true, "description": "URL of the next page, or null." },
          "previous": { "type": "string", "nullable": true, "description": "URL of the previous page, or null." },
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/BatchJob" } }
        },
        "required": ["count", "results"]
      },
...
    }
  }
}